// src/interfaces/dashboard.interface.ts
export interface IDashboardDataResponse {
  campaignCount: string;
  activeCampaignCount: string;
  totalDonationRaised: string;
  donationCount: string;
}

export interface IAdminDashboardDataResponse {
  campaignCount: string;
  donationAmount: string;
  totalUser: string;
  pendingKyc: string;
}